requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Werkzeug==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
flask-login==0.5.0
PyJWT==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.1

